+++ changelog.md
@@ -14,6 +14,7 @@
 - Biorthonormalization works for a set of close states as well. eT-program/eT!775
 - Added printing of implementation references, and the associated keyword `full references` in the `print` section. eT-program/eT!773
 - Made autogenerate script compatible with Python 3.6 (UTF-8 encoding error). eT-program/eT!794
+- Added `ghost` atoms. eT-prgram/eT!795
 
 ### Optimization
 - One-electron integrals with screening based on the overlap matrix. eT-program/eT!770